Add Transaction Activity
Working on activity for adding transactions. Most of layout figured out. Need to figure out how to choose date using date picker, so user can select date on calendar instead of typing in date. Also, need to develop a spinner that will display envelopes, since it will be a database query to fill the choices for the spinner.