Arithmetic and Dates
Established Arithmetic when adding Envelopes and Transactions. Added ID to Transaction model as Primary Key and using another Realm Object called Globals to keep tabs on the ID for Transactions and overall balance of all of the Envelopes. Dates are displaying properly and when adding a transaction the current date is set as default.